chore: refactor scan.py into modular architecture for maintainability
Decompose monolithic scan.py (1,891 lines) into clean modular architecture:

Architecture Changes:
- Main entry point: scan.py reduced from 1,891 to 353 lines
- Created scan_core package with clear separation of concerns
- All modules under 500-line limit (largest: 245 lines)
- Business logic functions refactored for clarity (e.g., advisory
  parser split into 4 focused schema handlers)

Module Organization:
- Core: models, config, utils, advisory, scanner
- Scanners: lockfiles, package_json, iocs, patterns, exfiltration,
  cache, node_modules
- Reporting: formatters, structured, compact, json_output

Quality Improvements:
- Eliminated code duplication via helper functions
- Reduced complexity in business logic (advisory.py depth 4→2)
- Maintained backward compatibility via re-exports
- All tests passing (81% coverage)
- Zero linting errors

Documentation:
- Updated CLAUDE.md with pragmatic quality standards
- Emphasized KISS/YAGNI principles over rigid metrics
- Added guidance on maintainability vs metric compliance

Closes #17
