+++ scripts/scan_core/__init__.py
@@ -1,0 +1,4 @@
+"""Shai-Hulud scan core modules."""
+from scripts.scan_core.models import Finding, ScanStats
+
+__all__ = ["Finding", "ScanStats"]
