+++ scripts/scan_core/reporting/__init__.py
@@ -1,0 +1,1 @@
+"""Reporting modules for scan output formatting."""
