+++ scripts/scan_core/scanners/__init__.py
@@ -1,0 +1,1 @@
+"""Scanner modules for different file types and detection methods."""
